feat(backuplogs): shorten the start/end date format. Fixes #1535

+++ src/xo-app/logs/index.js
@@ -104,13 +104,13 @@ const LOG_COLUMNS = [
   },
   {
     name: _('jobStart'),
-    itemRenderer: log => log.start && <FormattedDate value={new Date(log.start)} month='long' day='numeric' year='numeric' hour='2-digit' minute='2-digit' second='2-digit' />,
+    itemRenderer: log => log.start && <FormattedDate value={new Date(log.start)} month='short' day='numeric' year='numeric' hour='2-digit' minute='2-digit' second='2-digit' />,
     sortCriteria: log => log.start,
     sortOrder: 'desc'
   },
   {
     name: _('jobEnd'),
-    itemRenderer: log => log.end && <FormattedDate value={new Date(log.end)} month='long' day='numeric' year='numeric' hour='2-digit' minute='2-digit' second='2-digit' />,
+    itemRenderer: log => log.end && <FormattedDate value={new Date(log.end)} month='short' day='numeric' year='numeric' hour='2-digit' minute='2-digit' second='2-digit' />,
     sortCriteria: log => log.end,
     sortOrder: 'desc'
   },
